Add initial config implementation.
There are 3 parts of the configuration, initialization and
usage.

Validation is handled by jsonschema, small bits and pieces
are added into the corresponding dataclasses but nothing too important.

To handle initialization a simple pattern of what logging or django do has
been choes. Force user to call a certain 'initialization' funciton at
least once in the lifetime of the interpreter before making the config itself
usable.

By doing this we ensure that chached config is not going to change
during the whole execution and that only one instance (per interpreter)
of it will exist in memory.

While this is adds a little bit more on the user, it greatly simplifies
testing.

As for usage, dataclasses and OO